@@ -1,6 +1,7 @@
+#source 'https://rubygems.org'
 source 'https://rubygems.org'
-ruby "2.0.0"
 
+ruby "2.1.1"
 gem 'rails', '4.0.2'
 
 # Bundle edge Rails instead:
@@ -8,20 +9,21 @@ gem 'rails', '4.0.2'
 
 gem 'mongoid', github: 'mongoid/mongoid'
 gem 'pusher'
-gem 'devise'
+gem 'devise', '3.1.0'
 gem "haml", ">= 3.0.0"
 gem "haml-rails"
 gem 'cancan'
 gem 'simple_form'
 gem 'bootstrap-sass', '~> 2.3.0.1'
 gem 'newrelic_rpm'
-gem 'websocket-rails'
+#gem 'websocket-rails'
 gem 'protected_attributes'
 
 gem 'bson'
 gem 'moped', github: "mongoid/moped"
 gem 'omniauth'
 gem 'omniauth-twitter', :github => 'arunagw/omniauth-twitter'
+gem "github_api"
 
 group :development do
   gem "better_errors"
@@ -43,7 +45,6 @@ gem 'uglifier'#, '>= 1.0.3'
 gem 'carrierwave-mongoid', :require => 'carrierwave/mongoid'#, github: 'carrierwaveuploader/carrierwave-mongoid'
 gem 'mongoid-grid_fs', github: 'ahoward/mongoid-grid_fs'
 gem "jquery-fileupload-rails"
-gem 'capistrano', '~> 2.15'
 gem 'quiet_assets', group: :development
 
 gem 'jquery-rails'
@@ -57,8 +58,23 @@ gem 'jquery-rails'
 # Use unicorn as the app server
 # gem 'unicorn'
 
-# Deploy with Capistrano
-# gem 'capistrano'
+# Deploy with Mina
+ gem 'mina'
 
 # To use debugger
 # gem 'debugger'
+
+# Gemfile
+group :development, :test do
+  gem 'factory_girl_rails'
+  gem 'minitest-rails'
+  gem 'minitest-matchers'
+  gem 'm'
+  gem 'mocha'
+  gem 'database_cleaner'
+  gem 'valid_attribute'
+  #gem 'minitest-metadata'
+  gem 'minitest-implicit-subject'
+  gem 'minitest-spec-expect'
+  gem 'minitest-rails-capybara'
+end
