@@ -6,7 +6,7 @@ RailsRumble::Application.routes.draw do
     end
   end
 
-  devise_for :users, :controllers => {:omniauth_callbacks => 'users/omniauth_callbacks'}
+  devise_for :users, :controllers => {:omniauth_callbacks => 'users/omniauth_callbacks', sessions: "users/sessions"}
   resources :games do
     collection do
       match 'practise', via: :get
