@@ -4,7 +4,9 @@
 #  Pusher.secret = 'f4faf47f16352976d518' 
 #else
   Pusher.app_id = '57169'
-  Pusher.key = '0549d325d1840b3a4c05'
-  Pusher.secret = '695b40954fec56afd574'   
+  Pusher.key = 'e37f26571306fa33919de3a32cdc7a5d'
+  Pusher.secret = '40cb7d5fb4f0c23d30b3b835235b1921'   
 #end 
+Pusher.host = '54.186.39.143'
+Pusher.port = 4567
 
